+++ attributes/default.rb
@@ -38,4 +38,5 @@ default['user']['manage_home']        = "true"
 default['user']['create_user_group']  = "true"
 default['user']['ssh_keygen']         = "true"
 
-default['user']['data_bag'] = "users"
+default['user']['data_bag']             = "users"
+default['user']['user_array_node_attr'] = "users"